Enhances authorization and refines initial data seeding
Refactors the authorization middleware to support granular, role-based, and owner-specific permissions. Introduces an "admin:manage" override for comprehensive access control.

Improves data seeding for permissions and roles by making seeders idempotent, preventing re-creation on server restarts. Roles are now dynamically linked to permissions using `ObjectId` references, and the default user role assignment fetches the "user" role's ID programmatically.

Updates the `roles` schema to store permission references as `ObjectId`s, improving data integrity. Adds `id` validation to the user update schema and includes new entries in `.gitignore`.

+++ src/seeders/permission.seeder.js
@@ -1,46 +1,43 @@
-import { faker } from "@faker-js/faker";
 import Permission from "../db/models/permission.model.js";
+import { asyncHandler } from "../utils/asyncHandler.js";
 
-const permissionSeed = async() => {
-  try {
+const permissions = [
+  {
+    name: "admin:manage",
+  },
+  {
+    name: "user:update_own",
+  },
+  {
+    name: "user:delete_own",
+  },
+  {
+    name: "user:show_own",
+  },
+  {
+    name: "user:show_all",
+  },
+  {
+    name: "user:create",
+  },
+  {
+    name: "user:update",
+  },
+  {
+    name: "user:replace",
+  },
+  {
+    name: "user:delete",
+  },
+];
+
+const permissionSeed = asyncHandler(async () => {
+  const permissionCount = await Permission.countDocuments({});
+  if (permissionCount === 0) {
     await Permission.deleteMany({});
-    const permissions = [
-      {
-        name: "user:manage",
-      },
-      {
-        name: "user:update_own",
-      },
-      {
-        name: "user:delete_own"
-      },
-      {
-        name: "user:show_own"
-      },
-      {
-        name: "user:show_all"
-      },
-      {
-        name: "user:create"
-      },
-      {
-        name: "user:update"
-      },
-      {
-        name: "user:replace",
-      },
-      {
-        name: "user:delete"
-      }
-    ];
-
-    await Permission.insertMany(permissions)
-
-    console.log("permissions collection has been updated")
-
-  } catch (err) {
-    console.log(err);
+    const addedPermissions = await Permission.insertMany(permissions);
+    console.log("permissions has been initalized");
   }
-};
+});
 
-export default permissionSeed
+export default permissionSeed;
